Replacing backslashes for forward in windows paths
I have moved the spot where blackslashes are swapped for forward.

The old behaviour was checking for windows style paths by checking
the start of the path, but since at that point the front part would
be swapped for a unix style path it wasn't being detected as a win
style path.

Now we use the is_win which should be more reliable, and the
replacement happens inside the existing if statment.